requirements.txt
Django==1.7.8
django-admin-bootstrapped==2.4.0
django-filter==1.0.1
djangorestframework==3.3.1
Markdown==2.6.7
MySQL-python==1.2.5
Pillow==3.4.2
psycopg2==2.6.2
redis==2.10.5
